fixed file structure and message container
